# server/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
)

// AddRestaurantRequest defines model for AddRestaurantRequest.
type AddRestaurantRequest struct {
	Address       string  `json:"address"`
	Categories    []Uuid  `json:"categories"`
	Closed        bool    `json:"closed"`
	GooglePlaceId string  `json:"googlePlaceId"`
	Lat           float64 `json:"lat"`
	Lng           float64 `json:"lng"`
	Name          string  `json:"name"`
	PostalCode    string  `json:"postalCode"`
}

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	User User `json:"user"`
}

// CategoriesResponse defines model for CategoriesResponse.
type CategoriesResponse struct {
	Categories []Category `json:"categories"`
}

// Category defines model for Category.
type Category struct {
	Icon string `json:"icon"`

	// Id UUID serialized as a string.
	Id    Uuid   `json:"id"`
	Label string `json:"label"`
}

// ErrorBody defines model for ErrorBody.
type ErrorBody struct {
	Error string `json:"error"`
}

// GoogleAuthRequest defines model for GoogleAuthRequest.
type GoogleAuthRequest struct {
	Credential string `json:"credential"`
}

// Restaurant defines model for Restaurant.
type Restaurant struct {
	Address       string `json:"address"`
	Categories    []Uuid `json:"categories"`
	Closed        bool   `json:"closed"`
	Favorite      bool   `json:"favorite"`
	GooglePlaceId string `json:"googlePlaceId"`

	// Id UUID serialized as a string.
	Id         Uuid    `json:"id"`
	Lat        float64 `json:"lat"`
	Lng        float64 `json:"lng"`
	Name       string  `json:"name"`
	PostalCode string  `json:"postalCode"`
	Rate       float64 `json:"rate"`
	Visited    bool    `json:"visited"`
}

// RestaurantsResponse defines model for RestaurantsResponse.
type RestaurantsResponse struct {
	Restaurants []Restaurant `json:"restaurants"`
}

// UpdateRestaurantRequest defines model for UpdateRestaurantRequest.
type UpdateRestaurantRequest struct {
	Address       string  `json:"address"`
	Closed        bool    `json:"closed"`
	Favorite      bool    `json:"favorite"`
	GooglePlaceId string  `json:"googlePlaceId"`
	Lat           float64 `json:"lat"`
	Lng           float64 `json:"lng"`
	Name          string  `json:"name"`
	PostalCode    string  `json:"postalCode"`
	Rate          float64 `json:"rate"`
	Visited       bool    `json:"visited"`
}

// User defines model for User.
type User struct {
	Email string `json:"email"`

	// Id UUID serialized as a string.
	Id      Uuid `json:"id"`
	IsAdmin bool `json:"isAdmin"`
}

// Uuid UUID serialized as a string.
type Uuid = string

// GoogleAuthJSONRequestBody defines body for GoogleAuth for application/json ContentType.
type GoogleAuthJSONRequestBody = GoogleAuthRequest

// AddRestaurantJSONRequestBody defines body for AddRestaurant for application/json ContentType.
type AddRestaurantJSONRequestBody = AddRestaurantRequest

// UpdateRestaurantJSONRequestBody defines body for UpdateRestaurant for application/json ContentType.
type UpdateRestaurantJSONRequestBody = UpdateRestaurantRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/v1/auth/google)
	GoogleAuth(c *gin.Context)

	// (POST /api/v1/auth/logout)
	Logout(c *gin.Context)

	// (GET /api/v1/auth/me)
	Me(c *gin.Context)

	// (POST /api/v1/auth/restaurants)
	AddRestaurant(c *gin.Context)

	// (PUT /api/v1/auth/restaurants/{id})
	UpdateRestaurant(c *gin.Context, id Uuid)

	// (GET /api/v1/categories)
	ListCategories(c *gin.Context)

	// (GET /api/v1/restaurants)
	ListRestaurants(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GoogleAuth operation middleware
func (siw *ServerInterfaceWrapper) GoogleAuth(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GoogleAuth(c)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Logout(c)
}

// Me operation middleware
func (siw *ServerInterfaceWrapper) Me(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Me(c)
}

// AddRestaurant operation middleware
func (siw *ServerInterfaceWrapper) AddRestaurant(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AddRestaurant(c)
}

// UpdateRestaurant operation middleware
func (siw *ServerInterfaceWrapper) UpdateRestaurant(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Uuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateRestaurant(c, id)
}

// ListCategories operation middleware
func (siw *ServerInterfaceWrapper) ListCategories(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListCategories(c)
}

// ListRestaurants operation middleware
func (siw *ServerInterfaceWrapper) ListRestaurants(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListRestaurants(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/api/v1/auth/google", wrapper.GoogleAuth)
	router.POST(options.BaseURL+"/api/v1/auth/logout", wrapper.Logout)
	router.GET(options.BaseURL+"/api/v1/auth/me", wrapper.Me)
	router.POST(options.BaseURL+"/api/v1/auth/restaurants", wrapper.AddRestaurant)
	router.PUT(options.BaseURL+"/api/v1/auth/restaurants/:id", wrapper.UpdateRestaurant)
	router.GET(options.BaseURL+"/api/v1/categories", wrapper.ListCategories)
	router.GET(options.BaseURL+"/api/v1/restaurants", wrapper.ListRestaurants)
}

type GoogleAuthRequestObject struct {
	Body *GoogleAuthJSONRequestBody
}

type GoogleAuthResponseObject interface {
	VisitGoogleAuthResponse(w http.ResponseWriter) error
}

type GoogleAuth200JSONResponse AuthResponse

func (response GoogleAuth200JSONResponse) VisitGoogleAuthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GoogleAuth400JSONResponse ErrorBody

func (response GoogleAuth400JSONResponse) VisitGoogleAuthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GoogleAuth500JSONResponse ErrorBody

func (response GoogleAuth500JSONResponse) VisitGoogleAuthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type LogoutRequestObject struct {
}

type LogoutResponseObject interface {
	VisitLogoutResponse(w http.ResponseWriter) error
}

type Logout200Response struct {
}

func (response Logout200Response) VisitLogoutResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type Logout400JSONResponse ErrorBody

func (response Logout400JSONResponse) VisitLogoutResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Logout401JSONResponse ErrorBody

func (response Logout401JSONResponse) VisitLogoutResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type Logout500JSONResponse ErrorBody

func (response Logout500JSONResponse) VisitLogoutResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type MeRequestObject struct {
}

type MeResponseObject interface {
	VisitMeResponse(w http.ResponseWriter) error
}

type Me200JSONResponse User

func (response Me200JSONResponse) VisitMeResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type Me400JSONResponse ErrorBody

func (response Me400JSONResponse) VisitMeResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Me401JSONResponse ErrorBody

func (response Me401JSONResponse) VisitMeResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type Me500JSONResponse ErrorBody

func (response Me500JSONResponse) VisitMeResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type AddRestaurantRequestObject struct {
	Body *AddRestaurantJSONRequestBody
}

type AddRestaurantResponseObject interface {
	VisitAddRestaurantResponse(w http.ResponseWriter) error
}

type AddRestaurant200JSONResponse Restaurant

func (response AddRestaurant200JSONResponse) VisitAddRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type AddRestaurant400JSONResponse ErrorBody

func (response AddRestaurant400JSONResponse) VisitAddRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type AddRestaurant401JSONResponse ErrorBody

func (response AddRestaurant401JSONResponse) VisitAddRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type AddRestaurant500JSONResponse ErrorBody

func (response AddRestaurant500JSONResponse) VisitAddRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type UpdateRestaurantRequestObject struct {
	Id   Uuid `json:"id"`
	Body *UpdateRestaurantJSONRequestBody
}

type UpdateRestaurantResponseObject interface {
	VisitUpdateRestaurantResponse(w http.ResponseWriter) error
}

type UpdateRestaurant200Response struct {
}

func (response UpdateRestaurant200Response) VisitUpdateRestaurantResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type UpdateRestaurant400JSONResponse ErrorBody

func (response UpdateRestaurant400JSONResponse) VisitUpdateRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type UpdateRestaurant401JSONResponse ErrorBody

func (response UpdateRestaurant401JSONResponse) VisitUpdateRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type UpdateRestaurant500JSONResponse ErrorBody

func (response UpdateRestaurant500JSONResponse) VisitUpdateRestaurantResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ListCategoriesRequestObject struct {
}

type ListCategoriesResponseObject interface {
	VisitListCategoriesResponse(w http.ResponseWriter) error
}

type ListCategories200JSONResponse CategoriesResponse

func (response ListCategories200JSONResponse) VisitListCategoriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListCategories500JSONResponse ErrorBody

func (response ListCategories500JSONResponse) VisitListCategoriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ListRestaurantsRequestObject struct {
}

type ListRestaurantsResponseObject interface {
	VisitListRestaurantsResponse(w http.ResponseWriter) error
}

type ListRestaurants200JSONResponse RestaurantsResponse

func (response ListRestaurants200JSONResponse) VisitListRestaurantsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListRestaurants500JSONResponse ErrorBody

func (response ListRestaurants500JSONResponse) VisitListRestaurantsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (POST /api/v1/auth/google)
	GoogleAuth(ctx context.Context, request GoogleAuthRequestObject) (GoogleAuthResponseObject, error)

	// (POST /api/v1/auth/logout)
	Logout(ctx context.Context, request LogoutRequestObject) (LogoutResponseObject, error)

	// (GET /api/v1/auth/me)
	Me(ctx context.Context, request MeRequestObject) (MeResponseObject, error)

	// (POST /api/v1/auth/restaurants)
	AddRestaurant(ctx context.Context, request AddRestaurantRequestObject) (AddRestaurantResponseObject, error)

	// (PUT /api/v1/auth/restaurants/{id})
	UpdateRestaurant(ctx context.Context, request UpdateRestaurantRequestObject) (UpdateRestaurantResponseObject, error)

	// (GET /api/v1/categories)
	ListCategories(ctx context.Context, request ListCategoriesRequestObject) (ListCategoriesResponseObject, error)

	// (GET /api/v1/restaurants)
	ListRestaurants(ctx context.Context, request ListRestaurantsRequestObject) (ListRestaurantsResponseObject, error)
}

type StrictHandlerFunc = strictgin.StrictGinHandlerFunc
type StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GoogleAuth operation middleware
func (sh *strictHandler) GoogleAuth(ctx *gin.Context) {
	var request GoogleAuthRequestObject

	var body GoogleAuthJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GoogleAuth(ctx, request.(GoogleAuthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GoogleAuth")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GoogleAuthResponseObject); ok {
		if err := validResponse.VisitGoogleAuthResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// Logout operation middleware
func (sh *strictHandler) Logout(ctx *gin.Context) {
	var request LogoutRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Logout(ctx, request.(LogoutRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Logout")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(LogoutResponseObject); ok {
		if err := validResponse.VisitLogoutResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// Me operation middleware
func (sh *strictHandler) Me(ctx *gin.Context) {
	var request MeRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Me(ctx, request.(MeRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Me")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(MeResponseObject); ok {
		if err := validResponse.VisitMeResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// AddRestaurant operation middleware
func (sh *strictHandler) AddRestaurant(ctx *gin.Context) {
	var request AddRestaurantRequestObject

	var body AddRestaurantJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.AddRestaurant(ctx, request.(AddRestaurantRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "AddRestaurant")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(AddRestaurantResponseObject); ok {
		if err := validResponse.VisitAddRestaurantResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// UpdateRestaurant operation middleware
func (sh *strictHandler) UpdateRestaurant(ctx *gin.Context, id Uuid) {
	var request UpdateRestaurantRequestObject

	request.Id = id

	var body UpdateRestaurantJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateRestaurant(ctx, request.(UpdateRestaurantRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateRestaurant")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(UpdateRestaurantResponseObject); ok {
		if err := validResponse.VisitUpdateRestaurantResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListCategories operation middleware
func (sh *strictHandler) ListCategories(ctx *gin.Context) {
	var request ListCategoriesRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ListCategories(ctx, request.(ListCategoriesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListCategories")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(ListCategoriesResponseObject); ok {
		if err := validResponse.VisitListCategoriesResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListRestaurants operation middleware
func (sh *strictHandler) ListRestaurants(ctx *gin.Context) {
	var request ListRestaurantsRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ListRestaurants(ctx, request.(ListRestaurantsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListRestaurants")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(ListRestaurantsResponseObject); ok {
		if err := validResponse.VisitListRestaurantsResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}